setup news activity and layout, edit api and response classes, edit iview contracts

+++ app/src/main/java/artemshumidub/ru/sebbianews/data/repository/IRemoteRepositoryContract.java
@@ -9,7 +9,7 @@ import io.reactivex.Observable;
 public interface IRemoteRepositoryContract  {
 
     Observable<CategoryResponse> getCategory();
-    Observable<NewsListByCategoryResponse> getNewsList(int categoryId, int page);
-    Observable<NewsResponse> getNews(int ibNews);
+    Observable<NewsListByCategoryResponse> getNewsList(long categoryId, int page);
+    Observable<NewsResponse> getNews(long ibNews);
 
 }
